requirements.txt
# Core Agent Framework
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# Database Connectors
snowflake-connector-python>=3.6.0

# Configuration & Environment
python-dotenv>=1.0.0

# MCP Integration
mcp>=0.9.0

# Memory & Storage
sqlalchemy>=2.0.0

# Utilities
tabulate>=0.9.0
